Baby - Add/Edit/Delete by API

+++ app/Http/Controllers/Api/APIBabiesController.php
@@ -5,6 +5,7 @@ use Illuminate\Http\Request;
 use App\Http\Transformers\BabiesTransformer;
 use App\Http\Controllers\Api\BaseApiController;
 use App\Repositories\Babies\EloquentBabiesRepository;
+use Illuminate\Support\Facades\Validator;
 
 class APIBabiesController extends BaseApiController
 {
@@ -74,7 +75,24 @@ class APIBabiesController extends BaseApiController
     public function create(Request $request)
     {
         $input = $request->all();
-        $input = array_merge($input, ['image' => 'default.png']);
+
+        $validator = Validator::make($request->all(), [
+            'title'     => 'required'
+        ]);
+
+        if($validator->fails()) 
+        {
+            $messageData = '';
+
+            foreach($validator->messages()->toArray() as $message)
+            {
+                $messageData = $message[0];
+            }
+            return $this->failureResponse($validator->messages(), $messageData);
+        }
+
+        $userInfo   = $this->getAuthenticatedUser();
+        $input      = array_merge($input, ['image' => 'default.png', 'parent_id' => $userInfo->id]);
 
         if($request->file('image'))
         {
@@ -132,18 +150,46 @@ class APIBabiesController extends BaseApiController
      */
     public function edit(Request $request)
     {
-        $itemId = (int) hasher()->decode($request->get($this->primaryKey));
+        $input = $request->all();
+
+        $validator = Validator::make($request->all(), [
+            'title'     => 'required',
+            'baby_id'   => 'required',
+        ]);
+
+        if($validator->fails()) 
+        {
+            $messageData = '';
+
+            foreach($validator->messages()->toArray() as $message)
+            {
+                $messageData = $message[0];
+            }
+            return $this->failureResponse($validator->messages(), $messageData);
+        }
+
+        $userInfo   = $this->getAuthenticatedUser();
+        
+        if($request->file('image'))
+        {
+            $imageName  = rand(11111, 99999) . '_baby.' . $request->file('image')->getClientOriginalExtension();
+            $request->file('image')->move(base_path() . '/public/uploads/babies/', $imageName);
+            $input = array_merge($input, ['image' => $imageName]);
+        }
+
+        $itemId = (int) hasher()->decode($request->get('baby_id'));
+
 
         if($itemId)
         {
-            $status = $this->repository->update($itemId, $request->all());
+            $status = $this->repository->update($itemId, $input);
 
             if($status)
             {
                 $itemData       = $this->repository->getById($itemId);
                 $responseData   = $this->babiesTransformer->transform($itemData);
 
-                return $this->successResponse($responseData, 'Babies is Edited Successfully');
+                return $this->successResponse($responseData, 'Baby is Edited Successfully');
             }
         }
 
@@ -160,22 +206,33 @@ class APIBabiesController extends BaseApiController
      */
     public function delete(Request $request)
     {
-        $itemId = (int) hasher()->decode($request->get($this->primaryKey));
+        $itemId = (int) hasher()->decode($request->get('baby_id'));
 
         if($itemId)
         {
-            $status = $this->repository->destroy($itemId);
-
-            if($status)
+            
+            $userInfo   = $this->getAuthenticatedUser();
+            $babyCount  = $this->repository->model->where([
+                'id'        => $request->get('baby_id'),
+                'parent_id' => $userInfo->id
+                ])->count();
+            
+            if($babyCount > 0)
             {
-                return $this->successResponse([
-                    'success' => 'Babies Deleted'
-                ], 'Babies is Deleted Successfully');
+                $status     = $this->repository->destroy($itemId);
+
+                if($status)
+                {
+                    return $this->successResponse([
+                        'success' => 'Baby Deleted'
+                    ], 'Baby is Deleted Successfully');
+                }
             }
+
         }
 
         return $this->setStatusCode(404)->failureResponse([
             'reason' => 'Invalid Inputs'
-        ], 'Something went wrong !');
+        ], 'Baby Not Found or Baby Already Deleted !');
     }
 }
